py/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.10.0
websockets>=12.0
python-telegram-bot>=21.0
discord.py>=2.4.0
